factor liked dishes into separate file
